Moved MetricSelections to metricTypeComponent since it's not used outside of it
