+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -165,13 +165,6 @@ export enum AttributeTypeValue {
 	relative = "relative"
 }
 
-export enum MetricSelections {
-	areaMetric = "areaMetric",
-	heightMetric = "heightMetric",
-	colorMetric = "colorMetric",
-	edgeMetric = "edgeMetric"
-}
-
 export interface Edge {
 	fromNodeName: string
 	toNodeName: string
