+++ visualization/app/codeCharta/ui/metricType/metricType.component.spec.ts
@@ -1,9 +1,9 @@
 import "./metricType.module"
-import { MetricTypeController } from "./metricType.component"
+import { MetricSelections, MetricTypeController } from "./metricType.component"
 import { getService, instantiateModule } from "../../../../mocks/ng.mockhelper"
 import { MetricService } from "../../state/metric.service"
 import { IRootScopeService } from "angular"
-import { AttributeTypeValue, MetricSelections } from "../../codeCharta.model"
+import { AttributeTypeValue } from "../../codeCharta.model"
 import { CodeMapBuilding } from "../codeMap/rendering/codeMapBuilding"
 import { CodeMapMouseEventService } from "../codeMap/codeMap.mouseEvent.service"
 import { StoreService } from "../../state/store.service"
